feat: aplicando contexto para a comunicação com o Countdown
