+++ ignite-reactjs-2023/03-ignite-timer/src/pages/Home/components/Countdown/index.tsx
@@ -1,6 +1,59 @@
+import { useContext, useEffect, useState } from "react";
 import { CountdownContainer, Separator } from "./styles";
+import { differenceInSeconds } from "date-fns";
+import { CyclesContext } from "../..";
 
 export function Countdown(){
+    const { activeCycle, activeCycleId, markCurrentCycleAsFinished } = useContext(CyclesContext);
+    const [amountSecondsPassed, setAmountSecondsPassed] = useState(0);
+    const totalSeconds = activeCycle ? activeCycle.minutesAmount * 60 : 0;
+
+    useEffect(() => {
+        let interval : number;
+        if(activeCycle){
+            interval = setInterval(() => {
+                const secondsDifference = differenceInSeconds(new Date(), activeCycle.startDate);
+
+                if(secondsDifference >= totalSeconds){
+                    markCurrentCycleAsFinished();
+                    // setCycles((state) =>
+                    //     state.map((cycle) => {
+                    //         if(cycle.id === activeCycleId){
+                    //             return {...cycle, finishedDate: new Date() };
+                    //         } else {
+                    //             return cycle;
+                    //         }
+                    //     }),
+                    // )
+                    setAmountSecondsPassed(totalSeconds);
+                    clearInterval(interval);
+                }else{
+                    setAmountSecondsPassed(secondsDifference);
+                }
+            }, 1000);
+        }
+
+        return () => {
+            clearInterval(interval);
+        }
+
+    }, [activeCycle, totalSeconds, activeCycleId, markCurrentCycleAsFinished]);
+
+    const currentSeconds = activeCycle ? totalSeconds - amountSecondsPassed : 0;
+
+    const minutesAmount = Math.floor(currentSeconds/60);
+    const secondsAmount = currentSeconds % 60;
+
+    const minutes = String(minutesAmount).padStart(2, "0");
+    const seconds = String(secondsAmount).padStart(2, "0");
+
+    useEffect(() => {
+        if(activeCycle){
+            document.title = `${minutes}:${seconds}` 
+        }
+    }, [minutes,seconds, activeCycle]);
+
+
     return (
         <CountdownContainer>
             <span>{minutes[0]}</span>
