+++ ignite-reactjs-2023/03-ignite-timer/src/pages/Home/components/NewCycleForm/index.tsx
@@ -1,6 +1,35 @@
+import { useForm } from "react-hook-form";
 import { FormContainer, MinutesAmountInput, TaskInput } from "./styles";
+import * as zod from "zod";
+import { zodResolver } from "@hookform/resolvers/zod";
+
+
+const newCycleFormValidationSchema = zod.object({
+    task: zod.string().min(1, "Informe a tarefa"),
+    minutesAmount: zod
+        .number()
+        .min(1, "O ciclo precisa ser de no mínimo 5 minutos")
+        .max(60, "O ciclo precisa ser de no máximo 60 minutos"),
+});
+
+//o uso da interface está certo, mas o diego mosstrou como fazer com o zod.
+// interface NewCycleFormDate{
+//     task: string;
+//     minutesAmount: number;
+// }
+
+//Aqui eu estou inferindo o tipo de NewCycleFormDate usando o schema do zod lá em cima
+type NewCycleFormData = zod.infer<typeof newCycleFormValidationSchema>;
 
 export function NewCycleForm(){
+    const { register, handleSubmit, watch, reset  } = useForm<NewCycleFormData>({
+        resolver: zodResolver(newCycleFormValidationSchema),
+        defaultValues: {
+            task: "",
+            minutesAmount: 0,
+        }
+    });
+
     return (
         <FormContainer>
             <label htmlFor="">Vou trabalhar em</label>
